requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
click==8.1.7
cloudinary==1.41.0
deprecation==2.1.0
exceptiongroup==1.2.2
Flask==3.1.0
frozenlist==1.5.0
gotrue==2.11.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
numpy==1.21.6
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
postgrest==0.18.0
propcache==0.2.1
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.0.6
six==1.17.0
sniffio==1.3.1
storage3==0.9.0
StrEnum==0.4.15
supabase==2.10.0
supafunc==0.7.0
tflite-runtime==2.11.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
